+++ src/main/java/Floristeria/Client/Main.java
@@ -1,0 +1,16 @@
+package Floristeria.Client;
+
+import Floristeria.Client.Floristeria;
+
+import java.util.Scanner;
+
+public class Main {
+    static Scanner sc = new Scanner(System.in);
+
+    public static void main(String[] args) {
+
+        App app = new App();
+        app.start();
+    }
+}
+
