+++ src/main/java/Floristeria/Client/Ticket.java
@@ -1,0 +1,4 @@
+package Floristeria.Client;
+
+public class Ticket {
+}
