--- src/main/java/Floristeria/Floristeria.java
@@ -1,15 +1,0 @@
-package Floristeria;
-
-import Floristeria.Products.Arbol;
-import Floristeria.Products.Decoracion;
-import Floristeria.Products.Flor;
-
-import java.util.List;
-
-public class Floristeria {
-    private String nombre;
-    private List<Arbol> StockArboles;
-    private List<Flor> StockFlores;
-    private List<Decoracion> StockDecoracion;
-    private double valorStock;
-}
