--- src/main/java/Floristeria/Ticket.java
@@ -1,4 +1,0 @@
-package Floristeria;
-
-public class Ticket {
-}
